[UPDATE] refactor routes
