+++ backend/src/modules/airdrop/routes/fileRoutes.ts
@@ -1,14 +1,14 @@
 import express, { Request, Response } from "express"
-import { PrismaClient } from "@prisma/client"
-import { isNull } from "util"
+import { verifyUser } from "../../backendService/middleware/verifyUser"
+import getAllFile from "./functions/getAllFile"
+import uploadFile from "./functions/uploadFile"
+import downloadFile from "./functions/downloadFile"
+import hideFile from "./functions/hideFile"
 
 //file manage module
 const path = require("path")
 const fs = require("fs")
 const multer = require("multer")
-
-//prisma
-const prisma = new PrismaClient()
 //config destionation here
 const storage = multer.diskStorage({
     destination: function (req: any, file: any, cb: any) {
@@ -28,207 +28,15 @@ const fileRoutes = express()
 fileRoutes.use(express.json())
 
 fileRoutes.get("/", async (req: Request, res: Response) => {
-    console.log(req.user?.userId)
+    res.send(" Welcome to airdrop file API")
 })
 
-fileRoutes.get("/getallfile", async (req: Request, res: Response) => {
-    //fetch everone type
-    try {
-        const isShow = await prisma.user_Show_File.findMany({
-            where: {
-                userId: req.user?.userId,
-            },
-            select: {
-                fileId: true,
-            },
-        })
-        const everyone = await prisma.file_Info.findMany({
-            where: {
-                AND: [
-                    {
-                        fileSender: {
-                            not: req.user?.userId,
-                        },
-                    },
-                    {
-                        AND: [
-                            {
-                                fileId: {
-                                    notIn: isShow.map((item) => item.fileId),
-                                },
-                            },
-                            {
-                                sendType: "Everyone",
-                            },
-                        ],
-                    },
-                ],
-            },
-        })
-        //result will add senderName
-        const result = []
-        for (const item of everyone) {
-            const user = await prisma.user_Profile.findFirst({
-                where: {
-                    userId: item.fileSender,
-                },
-                select: {
-                    fName: true,
-                    lName: true,
-                },
-            })
-            const senderId = item.fileSender
-            result.push({
-                ...item,
-                fileSender: user?.fName + " " + user?.lName,
-                senderId: senderId,
-            })
-        }
-        const lastResult = result.map((item) => {
-            return {
-                ...item,
-                comments: [],
-            }
-        })
-        res.json(lastResult)
-    } catch (err) {
-        console.log(err)
-    }
-})
+fileRoutes.get("/getallfile",verifyUser,getAllFile)
+fileRoutes.post("/upload",verifyUser, upload.array("files"), uploadFile)
+fileRoutes.get("/download/:type/:filename", downloadFile)
+fileRoutes.post("/hidefile", hideFile)
 
-fileRoutes.post("/upload", upload.array("files"), async (req: Request, res: Response) => {
-    console.log("Upload sucessful")
-    console.log(req.body)
-    const sender = await req.user?.userId
-    console.log(sender)
-    try {
-        const payload: any = []
-        ;(req.files as Array<Express.Multer.File>).map((item: any) => {
-            const newDate = new Date(req.body.expireDate)
-            payload.push({
-                fileName: item.originalname,
-                fileSender: sender,
-                sendType: req.body.type,
-                fileDesc: req.body.description,
-                fileExpired: newDate,
-            })
-        })
-        const fileUpload = await prisma.file_Info.createMany({
-            data: payload,
-        })
-        //handle multiple receiver
-        if (req.body.receiver != null && req.body.receiver != "everyone") {
-            if (req.body.type == "department") {
-                const receiverListId = []
-                for (const item of req.body.receiver) {
-                    const receiver = await prisma.user_Profile.findFirst({
-                        where: {
-                            AND: [
-                                {
-                                    fName: item.split(" ")[0],
-                                },
-                                {
-                                    lName: item.split(" ")[0],
-                                },
-                            ],
-                        },
-                        select: {
-                            userId: true,
-                        },
-                    })
-                    receiverListId.push(receiver?.userId)
-                }
-                // insert file access to department table
-            } else if (req.body.type == "community") {
-                const receiverListId = []
-                for (const item of req.body.receiver) {
-                    const receiver = await prisma.community.findFirst({
-                        where: {
-                            AND: [
-                                {
-                                    communityName: item,
-                                },
-                            ],
-                        },
-                        select: {
-                            communityId: true,
-                        },
-                    })
-                    receiverListId.push(receiver?.communityId)
-                }
-            } else if (req.body.type == "specific") {
-                if (req.body.type == "department") {
-                    const receiverListId = []
-                    for (const item of req.body.receiver) {
-                        const receiver = await prisma.user_Profile.findFirst({
-                            where: {
-                                AND: [
-                                    {
-                                        fName: item.split(" ")[0],
-                                    },
-                                    {
-                                        lName: item.split(" ")[0],
-                                    },
-                                ],
-                            },
-                            select: {
-                                userId: true,
-                            },
-                        })
-                        receiverListId.push(receiver?.userId)
-                    }
-                    // insert file access to specific table
-                }
-            }
-        }
-    } catch (err) {
-        console.log(err)
-    }
-})
-fileRoutes.get("/download/:type/:filename", (req: Request, res: Response) => {
-    const directoryPath = path.join(__dirname, "../files" + "/" + req.params.type)
-    res.download(directoryPath + "/" + req.params.filename)
-})
-
-fileRoutes.post("/hidefile", async (req: Request, res: Response) => {
-    const { fileId } = req.body
-    // console.log(req.body);
-    const payload: any = {
-        userId: req.user?.userId,
-        fileId: fileId,
-    }
-    const hide = await prisma.user_Show_File.create({
-        data: payload,
-    })
-    res.json("hide file sucessful")
-})
-
-//function that will delete expired file
-const deleteExpiredFile = async () => {
-    const today = new Date()
-    const expiredFile = await prisma.file_Info.findMany({
-        where: {
-            fileExpired: {
-                lt: today,
-            },
-        },
-    })
-    expiredFile.map(async (item) => {
-        const directoryPath = path.join(__dirname, "../files" + "/" + item.sendType)
-        fs.unlink(directoryPath + "/" + item.fileName, (err: any) => {
-            if (err) {
-                console.log(err)
-            }
-        })
-        await prisma.file_Info.delete({
-            where: {
-                fileId: item.fileId,
-            },
-        })
-    })
-}
-
-//check expired file every 30 min
-// setInterval(deleteExpiredFile, 1800000)
+//check expired file every 5 minutes
+// setInterval(deleteExpiredFile, 300000)
 
 export default fileRoutes
